requirements.txt
langchain
streamlit
python-dotenv
langchain-community
openai
faiss-cpu
PyMuPDF
tiktoken
pypdf
